my_trading_engine/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
tiingo>=0.14.0
matplotlib>=3.4.0
dagster>=1.0.0
python-dotenv>=0.19.0
requests>=2.26.0